# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type ContactSubmissionSubject =
  (typeof ContactSubmissionSubject)[keyof typeof ContactSubmissionSubject];

export const ContactSubmissionSubject = {
  general: "general",
  partner: "partner",
  volunteer: "volunteer",
  sponsor: "sponsor",
  media: "media",
  other: "other",
} as const;

export interface ContactSubmission {
  /**
   * @minLength 1
   * @maxLength 200
   */
  name: string;
  /** @maxLength 320 */
  email: string;
  /** @maxLength 200 */
  organization?: string;
  subject: ContactSubmissionSubject;
  /**
   * @minLength 1
   * @maxLength 5000
   */
  message: string;
}

export interface ContactSubmissionResponse {
  ok: boolean;
  id: string;
}

export interface ErrorResponse {
  message: string;
}

export interface DeleteResponse {
  ok: boolean;
  id: number;
}

export interface NewsItemInput {
  /**
   * @minLength 1
   * @maxLength 300
   */
  title: string;
  /**
   * @minLength 1
   * @maxLength 2000
   */
  description: string;
  /**
   * @minLength 1
   * @maxLength 50
   */
  badge: string;
  /**
   * @minLength 1
   * @maxLength 50
   */
  badgeColor: string;
  /**
   * @minLength 1
   * @maxLength 1000
   */
  imageUrl: string;
  publishedAt?: string;
}

export interface NewsItem {
  id: number;
  title: string;
  description: string;
  badge: string;
  badgeColor: string;
  imageUrl: string;
  publishedAt: string;
  createdAt: string;
}

export interface EventInput {
  /**
   * @minLength 1
   * @maxLength 300
   */
  title: string;
  /**
   * @minLength 1
   * @maxLength 50
   */
  category: string;
  /**
   * @minLength 1
   * @maxLength 300
   */
  location: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  time: string;
  startsAt: string;
  /** @maxLength 2000 */
  description?: string | null;
  featured?: boolean;
  /** @maxLength 1000 */
  registrationUrl?: string | null;
}

export interface Event {
  id: number;
  title: string;
  category: string;
  location: string;
  time: string;
  startsAt: string;
  description?: string | null;
  featured: boolean;
  registrationUrl?: string | null;
  createdAt: string;
}

export interface TeamMemberInput {
  /**
   * @minLength 1
   * @maxLength 200
   */
  name: string;
  /**
   * @minLength 1
   * @maxLength 200
   */
  role: string;
  /**
   * @minLength 1
   * @maxLength 5
   */
  initials: string;
  /** @maxLength 500 */
  linkedinUrl?: string | null;
  sortOrder?: number;
}

export interface TeamMember {
  id: number;
  name: string;
  role: string;
  initials: string;
  linkedinUrl?: string | null;
  sortOrder: number;
  createdAt: string;
}