# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: contact
    description: Contact form operations
  - name: news
    description: News article operations
  - name: events
    description: Event operations
  - name: team
    description: Team member operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /contact:
    post:
      operationId: submitContact
      tags: [contact]
      summary: Submit contact form
      description: Submit a contact form message from the website
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ContactSubmission"
      responses:
        "200":
          description: Submission accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactSubmissionResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /news:
    get:
      operationId: listNews
      tags: [news]
      summary: List news items
      description: Returns all news items, ordered by published date (newest first)
      responses:
        "200":
          description: News items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsItem"
    post:
      operationId: createNews
      tags: [news]
      summary: Create a news item (admin only)
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsItemInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /news/{id}:
    put:
      operationId: updateNews
      tags: [news]
      summary: Update a news item (admin only)
      security:
        - adminToken: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewsItemInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NewsItem"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteNews
      tags: [news]
      summary: Delete a news item (admin only)
      security:
        - adminToken: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      description: Returns all events, ordered by start date (soonest first)
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create an event (admin only)
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /events/{id}:
    put:
      operationId: updateEvent
      tags: [events]
      summary: Update an event (admin only)
      security:
        - adminToken: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event (admin only)
      security:
        - adminToken: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /team:
    get:
      operationId: listTeamMembers
      tags: [team]
      summary: List team members
      description: Returns all team members, ordered by sort order
      responses:
        "200":
          description: Team members
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamMember"
    post:
      operationId: createTeamMember
      tags: [team]
      summary: Create a team member (admin only)
      security:
        - adminToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /team/{id}:
    put:
      operationId: updateTeamMember
      tags: [team]
      summary: Update a team member (admin only)
      security:
        - adminToken: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamMemberInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamMember"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteTeamMember
      tags: [team]
      summary: Delete a team member (admin only)
      security:
        - adminToken: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  securitySchemes:
    adminToken:
      type: http
      scheme: bearer
      description: Bearer token for admin operations (set via ADMIN_TOKEN env var)
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    ContactSubmission:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        email:
          type: string
          format: email
          maxLength: 320
        organization:
          type: string
          maxLength: 200
        subject:
          type: string
          enum: [general, partner, volunteer, sponsor, media, other]
        message:
          type: string
          minLength: 1
          maxLength: 5000
      required:
        - name
        - email
        - subject
        - message
    ContactSubmissionResponse:
      type: object
      properties:
        ok:
          type: boolean
        id:
          type: string
      required:
        - ok
        - id
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
    DeleteResponse:
      type: object
      properties:
        ok:
          type: boolean
        id:
          type: integer
      required:
        - ok
        - id
    NewsItemInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 300
        description:
          type: string
          minLength: 1
          maxLength: 2000
        badge:
          type: string
          minLength: 1
          maxLength: 50
        badgeColor:
          type: string
          minLength: 1
          maxLength: 50
        imageUrl:
          type: string
          minLength: 1
          maxLength: 1000
        publishedAt:
          type: string
          format: date-time
      required:
        - title
        - description
        - badge
        - badgeColor
        - imageUrl
    NewsItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        badge:
          type: string
        badgeColor:
          type: string
        imageUrl:
          type: string
        publishedAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - description
        - badge
        - badgeColor
        - imageUrl
        - publishedAt
        - createdAt
    EventInput:
      type: object
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 300
        category:
          type: string
          minLength: 1
          maxLength: 50
        location:
          type: string
          minLength: 1
          maxLength: 300
        time:
          type: string
          minLength: 1
          maxLength: 100
        startsAt:
          type: string
          format: date-time
        description:
          type: string
          maxLength: 2000
          nullable: true
        featured:
          type: boolean
        registrationUrl:
          type: string
          maxLength: 1000
          nullable: true
      required:
        - title
        - category
        - location
        - time
        - startsAt
    Event:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        category:
          type: string
        location:
          type: string
        time:
          type: string
        startsAt:
          type: string
          format: date-time
        description:
          type: string
          nullable: true
        featured:
          type: boolean
        registrationUrl:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - category
        - location
        - time
        - startsAt
        - featured
        - createdAt
    TeamMemberInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
        role:
          type: string
          minLength: 1
          maxLength: 200
        initials:
          type: string
          minLength: 1
          maxLength: 5
        linkedinUrl:
          type: string
          maxLength: 500
          nullable: true
        sortOrder:
          type: integer
      required:
        - name
        - role
        - initials
    TeamMember:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        role:
          type: string
        initials:
          type: string
        linkedinUrl:
          type: string
          nullable: true
        sortOrder:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - role
        - initials
        - sortOrder
        - createdAt