# lib/api-zod/src/generated/types/event.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Event {
  id: number;
  title: string;
  category: string;
  location: string;
  time: string;
  startsAt: Date;
  description?: string | null;
  featured: boolean;
  registrationUrl?: string | null;
  createdAt: Date;
}