# lib/api-zod/src/generated/types/eventInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface EventInput {
  /**
   * @minLength 1
   * @maxLength 300
   */
  title: string;
  /**
   * @minLength 1
   * @maxLength 50
   */
  category: string;
  /**
   * @minLength 1
   * @maxLength 300
   */
  location: string;
  /**
   * @minLength 1
   * @maxLength 100
   */
  time: string;
  startsAt: Date;
  /** @maxLength 2000 */
  description?: string | null;
  featured?: boolean;
  /** @maxLength 1000 */
  registrationUrl?: string | null;
}