# lib/api-zod/src/generated/types/newsItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NewsItem {
  id: number;
  title: string;
  description: string;
  badge: string;
  badgeColor: string;
  imageUrl: string;
  publishedAt: Date;
  createdAt: Date;
}