# lib/api-zod/src/generated/types/newsItemInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface NewsItemInput {
  /**
   * @minLength 1
   * @maxLength 300
   */
  title: string;
  /**
   * @minLength 1
   * @maxLength 2000
   */
  description: string;
  /**
   * @minLength 1
   * @maxLength 50
   */
  badge: string;
  /**
   * @minLength 1
   * @maxLength 50
   */
  badgeColor: string;
  /**
   * @minLength 1
   * @maxLength 1000
   */
  imageUrl: string;
  publishedAt?: Date;
}