# lib/api-zod/src/generated/types/teamMember.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TeamMember {
  id: number;
  name: string;
  role: string;
  initials: string;
  linkedinUrl?: string | null;
  sortOrder: number;
  createdAt: Date;
}