# lib/api-zod/src/generated/types/teamMemberInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface TeamMemberInput {
  /**
   * @minLength 1
   * @maxLength 200
   */
  name: string;
  /**
   * @minLength 1
   * @maxLength 200
   */
  role: string;
  /**
   * @minLength 1
   * @maxLength 5
   */
  initials: string;
  /** @maxLength 500 */
  linkedinUrl?: string | null;
  sortOrder?: number;
}